binary-artifacts/requirements.txt
GitPython==3.1.12
boto3==1.16.58
Jinja2==2.11.2